requirements.txt
asgiref==3.8.1
bleach==6.2.0
boto3==1.38.36
botocore==1.38.36
dj-database-url==3.0.0
django==5.2.3
django-ckeditor==6.7.2
django-cleanup==9.0.0
django-js-asset==3.1.2
django-storages==1.14.6
django-summernote==0.8.20.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
whitenoise==6.9.0
